Add a simple CLI interface.
Prompts:

Anylyse this Python code. Respond only that you've analysed the code.

{code}

---

The name of this module is 'poker_hand`. It can be imported like this:

from poker_hand import PokerHand

---

Create a new Python module that use the Python Rich package to extend the functionality of PokerHand to a CLI interface.
It should be easy for the user to enter the value for 5 cards. If more or less values are entered, return a friendly error message.
The user will enter each value in the format "rank suit"

These are the valid rank values: {"ace", "2", "3", "4", "5", "6", "7", "8", "9", "10", "jack", "queen", "king"}
These are the valid suit values: {"hearts", "diamonds", "clubs", "spades"}