Remove duplicate test cases. (#7)
* Update documentation for Linting code.

* Formatting update from `isort`.

* Improve code quality.

Prompts:

Add Type Hints to this code

{code}

---

I have updated the code. Confirm that all Type Hints are defined for this code. If there are no correction to be made, reply with "Confirm". If there are missing Hints, show the changes required only.

{code}

* Improve code quality by adding DocStrings.

Prompts:

Create DocString comments for this module. Only display the DocStrings

{code}

---

I have updated the code. Confirm that all DocStrings are defined for this code. If there are no corrections to be made, reply with "Confirm". If there are missing DocStrings, fix the code and display it.

{code}

* Improve code quality by fixing pylint issue.

Prompts:

Fix this error in the code. Only display the fix.

C0103: Variable name "e" doesn't conform to snake_case naming style (invalid-name)

* Remove duplicate test data.